# internal/api/openapi.json
{
  "components": {
    "schemas": {
...
      "ApiError": {
        "additionalProperties": false,
        "properties": {
          "$schema": {
            "description": "A URL to the JSON Schema for this object.",
            "examples": [
              "https://example.com/ApiError.json"
            ],
            "format": "uri",
            "readOnly": true,
            "type": "string"
          },
          "error": {
            "$ref": "#/components/schemas/ApiErrorWrappedStruct"
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      },
      "ApiErrorWrappedStruct": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "type": "string"
          },
          "details": {},
          "message": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "type": "object"
      },
...
      "CallJSON": {
        "additionalProperties": false,
        "properties": {
          "$schema": {
            "description": "A URL to the JSON Schema for this object.",
            "examples": [
              "https://example.com/CallJSON.json"
            ],
            "format": "uri",
            "readOnly": true,
            "type": "string"
          },
          "answered_at": {
            "format": "date-time",
            "type": "string"
          },
          "direction": {
            "enum": [
              "in",
              "out"
            ],
            "type": "string"
          },
          "duration_ms": {
            "format": "int64",
            "type": "integer"
          },
          "end_reason": {
            "type": "string"
          },
          "ended_at": {
            "format": "date-time",
            "type": "string"
          },
          "from": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "started_at": {
            "format": "date-time",
            "type": "string"
          },
          "state": {
            "enum": [
              "ringing",
              "dialing",
              "alerting",
              "active",
              "held",
              "ended",
              "missed",
              "rejected"
            ],
            "type": "string"
          },
          "to": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "direction",
          "state",
          "started_at"
        ],
        "type": "object"
      },
...
      "CallsDialInputBody": {
        "additionalProperties": false,
        "properties": {
          "$schema": {
            "description": "A URL to the JSON Schema for this object.",
            "examples": [
              "https://example.com/CallsDialInputBody.json"
            ],
            "format": "uri",
            "readOnly": true,
            "type": "string"
          },
          "to": {
            "examples": [
              "+15551234567"
            ],
            "type": "string"
          }
        },
        "required": [
          "to"
        ],
        "type": "object"
      },
      "CallsListOutputBody": {
        "additionalProperties": false,
        "properties": {
          "$schema": {
            "description": "A URL to the JSON Schema for this object.",
            "examples": [
              "https://example.com/CallsListOutputBody.json"
            ],
            "format": "uri",
            "readOnly": true,
            "type": "string"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/CallJSON"
            },
            "type": [
              "array",
              "null"
            ]
          }
        },
        "required": [
          "items"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "opaque",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Cellular modem control plane: status, SMS, calls, events, admin.",
    "title": "SIM7600D Control API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/calls": {
      "get": {
        "operationId": "callsList",
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "explode": false,
            "in": "query",
            "name": "since",
            "schema": {
              "type": "string"
            }
          },
          {
            "explode": false,
            "in": "query",
            "name": "limit",
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CallsListOutputBody"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "List calls",
        "tags": [
          "calls"
        ]
      },
      "post": {
        "operationId": "callsDial",
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "schema": {
              "type": "string"
            }
          },
          {
            "explode": false,
            "in": "query",
            "name": "raw",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CallsDialInputBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CallJSON"
                }
              }
            },
            "description": "Created"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Place an outbound call",
        "tags": [
          "calls"
        ]
      }
    },
...
  }
}